# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Synesis Enterprise API
  version: 1.0.0
  description: API specification for the Synesis AI Talent Intelligence Platform.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enterprise/bulk-jobs:
    post:
      operationId: createBulkJob
      tags: [bulk-jobs]
      summary: Start a new bulk analysis job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobDescriptionId: { type: integer }
                resumeProfileIds:
                  type: array
                  items: { type: integer }
              required: [jobDescriptionId, resumeProfileIds]
      responses:
        "201":
          description: Job started
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...